Moved fav db.
