+++ app/src/main/java/com/projects/musicplayer/database/PlaylistRoomDatabase.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import androidx.room.TypeConverters
+import com.projects.musicplayer.rest.FavRoomDatabase
 
 
 //@TypeConverters(PlaylistConverter::class)
