+++ app/src/main/java/com/projects/musicplayer/rest/RecentSongConverter.kt
@@ -1,9 +1,6 @@
-package com.projects.musicplayer.database
+package com.projects.musicplayer.rest
 
 import androidx.room.TypeConverter
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
-import java.lang.reflect.Type
 import java.util.*
 
 //@ProvidedTypeConverter
